backend/requirements.txt
yfinance
pandas
requests
vaderSentiment
python-dotenv
schedule
newsapi-python
gnews
